Added except block for FileNotFoundError. (#27)
* Added except block for FileNotFoundError.

If a folder exists that indicates the presence of a repository (i.e
'.git', '.hg', and '.svn'), but the corresponding binary cannot be found
in the user's path, the application will now gracefully handle the error
by presenting the user with a message that the directory was skipped
because the corresponding directory could not be located.

Note that this works quite well for Mercurial and Subversion, but the
message is printed 3-5 times for Git (depending on which command line
options were provided). A more robust solution would check for the
presence of the Git binary _before_ calling 'run' in the status_git
function.