Remove file with out-of-date URL
The URL in this file caused pull request spam from a bot, so for them
moment I’m removing it.  If its user needed it, they will respond with a
pull request to reinstate it — but I myself don’t know the correct step
to keep it up to date.